resize image backend

+++ backend/src/main/kotlin/com/headblog/backend/domain/model/media/MediaSize.kt
@@ -2,5 +2,5 @@ package com.headblog.backend.domain.model.media
 
 data class MediaSize(
     val url: String,
-    val size: Long
+    val size: Int
 )
